Refactor email into separate logical components Remount existing functionality to fit revised API spec
