+++ app/src/routes/v1/email.js
@@ -3,8 +3,7 @@ const queueComponent = require('../../components/queue');
 
 const emailRouter = require('express').Router();
 const {
-  validateEmail,
-  validateMerge
+  validateEmail
 } = require('../../middleware/validation');
 
 /** Email sending endpoint */
@@ -19,31 +18,6 @@ emailRouter.post('/', validateEmail, async (req, res, next) => {
         messageId: result
       });
     }
-  } catch (error) {
-    next(error);
-  }
-});
-
-/** Template mail merge & email sending endpoint */
-emailRouter.post('/merge', validateMerge, async (req, res, next) => {
-  try {
-    if (req.query.devMode) {
-      const result = await emailComponent.mergeMailEthereal(req.body);
-      res.status(201).json(result);
-    } else {
-      const result = await emailComponent.mergeMailSmtp(req.body);
-      res.status(201).json(result);
-    }
-  } catch (error) {
-    next(error);
-  }
-});
-
-/** Template mail merge validation & preview endpoint */
-emailRouter.post('/merge/preview', validateMerge, (req, res, next) => {
-  try {
-    const result = emailComponent.mergeTemplate(req.body);
-    res.status(201).json(result);
   } catch (error) {
     next(error);
   }
