+++ app/src/routes/v1/health.js
@@ -1,9 +1,9 @@
 const checksRouter = require('express').Router();
 
-const checkComponent = require('../../components/checks');
+const checkComponent = require('../../components/health');
 
 /** Returns the status of correspondent APIs */
-checksRouter.get('/status', async (_req, res, next) => {
+checksRouter.get('/', async (_req, res, next) => {
   const statuses = await checkComponent.getStatus();
 
   if (statuses instanceof Array) {
