+++ app/tests/unit/components/email.spec.js
@@ -21,21 +21,6 @@ const subject = 'subject {{ foo }}';
 const url = 'https://example.com';
 
 // Object Fixtures
-const contextEntry = {
-  'to': [
-    'bar@example.com'
-  ],
-  'cc': [
-    'baz@example.com'
-  ],
-  'bcc': [
-    'foo@example.com',
-    'fizz@example.com'
-  ],
-  'context': {
-    'foo': 'test'
-  }
-};
 const envelope = {
   text: body,
   encoding: 'utf-8',
@@ -64,19 +49,6 @@ const message = {
   ],
   subject: subject
 };
-const template = {
-  attachments: undefined,
-  bodyType: 'text',
-  body: body,
-  'contexts': [
-    contextEntry,
-    contextEntry
-  ],
-  encoding: 'utf-8',
-  from: 'foo@example.com',
-  priority: 'normal',
-  subject: subject
-};
 
 describe('createEnvelope', () => {
   afterEach(() => {
@@ -90,98 +62,6 @@ describe('createEnvelope', () => {
     expect(result).toBeTruthy();
     expect(result.bodyType).toBeUndefined();
     expect(result.text).toMatch(body);
-  });
-});
-
-describe('mergeMailEthereal', () => {
-  let spy;
-
-  beforeEach(() => {
-    spy = jest.spyOn(email, 'sendMailEthereal');
-  });
-
-  afterEach(() => {
-    spy.mockRestore();
-  });
-
-  it('should yield an array of Ethereal email urls', async () => {
-    spy.mockResolvedValue(url);
-
-    const result = await email.mergeMailEthereal(template);
-    expect(result).toBeTruthy();
-    expect(result).toHaveLength(2);
-    expect(result[0]).toEqual(url);
-    expect(spy).toHaveBeenCalledTimes(2);
-  });
-
-  it('should throw an error if any sending failed', async () => {
-    spy.mockResolvedValueOnce(url);
-    spy.mockImplementation(() => {
-      throw new Error(errorMessage);
-    });
-
-    expect(email.mergeMailEthereal(template)).rejects.toThrow(errorMessage);
-  });
-});
-
-describe('mergeMailSmtp', () => {
-  let spy;
-
-  beforeEach(() => {
-    spy = jest.spyOn(email, 'sendMailSmtp');
-  });
-
-  afterEach(() => {
-    spy.mockRestore();
-  });
-
-  it('should yield an array of nodemailer result objects', async () => {
-    spy.mockResolvedValue(info);
-
-    const result = await email.mergeMailSmtp(template);
-    expect(result).toBeTruthy();
-    expect(result).toHaveLength(2);
-    expect(spy).toHaveBeenCalledTimes(2);
-  });
-
-  it('should throw an error if any sending failed', async () => {
-    spy.mockResolvedValueOnce(info);
-    spy.mockImplementation(() => {
-      throw new Error(errorMessage);
-    });
-
-    expect(email.mergeMailSmtp(template)).rejects.toThrow(errorMessage);
-  });
-});
-
-describe('mergeTemplate', () => {
-  it('should yield an array of message objects', () => {
-    const result = email.mergeTemplate(template);
-    expect(result).toBeTruthy();
-    expect(result).toHaveLength(2);
-    expect(result[0].body).toMatch('body test');
-    expect(result[0].to).toBeTruthy();
-    expect(result[0].subject).toMatch('subject test');
-  });
-});
-
-describe('renderMerge', () => {
-  const str = 'Hello {{ foo }}';
-  const context = {
-    foo: 'test'
-  };
-
-  it('should yield a rendered merge string', () => {
-    const result = email.renderMerge(str, context);
-    expect(result).toBeTruthy();
-    expect(result).toMatch('Hello test');
-  });
-
-  it('should throw an error on an unrecognized dialect', () => {
-    const dialect = 'badDialect';
-    const result = () => email.renderMerge(str, context, dialect);
-    expect(result).toBeTruthy();
-    expect(result).toThrow(`Dialect ${dialect} not supported`);
   });
 });
