+++ app/tests/unit/routes/v1/health.spec.js
@@ -1,11 +1,11 @@
 const express = require('express');
 const request = require('supertest');
 
-const router = require('../../../../src/routes/v1/checks');
-const checkComponent = require('../../../../src/components/checks');
+const router = require('../../../../src/routes/v1/health');
+const checkComponent = require('../../../../src/components/health');
 
 // Simple Express Server
-const basePath = '/api/v1/checks';
+const basePath = '/api/v1/health';
 const app = express();
 app.use(express.json());
 app.use(express.urlencoded({
@@ -13,7 +13,7 @@ app.use(express.urlencoded({
 }));
 app.use(basePath, router);
 
-describe(`GET ${basePath}/status`, () => {
+describe(`GET ${basePath}`, () => {
   afterEach(() => {
     checkComponent.getStatus.mockReset();
   });
@@ -21,7 +21,7 @@ describe(`GET ${basePath}/status`, () => {
   it('should return the status of correspondent apis', async () => {
     checkComponent.getStatus = jest.fn().mockResolvedValue([{}]);
 
-    const response = await request(app).get(`${basePath}/status`);
+    const response = await request(app).get(`${basePath}`);
 
     expect(response.statusCode).toBe(200);
     expect(response.body).toBeTruthy();
@@ -31,7 +31,7 @@ describe(`GET ${basePath}/status`, () => {
   it('should respond even with an exception', async () => {
     checkComponent.getStatus = jest.fn().mockResolvedValue({});
 
-    const response = await request(app).get(`${basePath}/status`);
+    const response = await request(app).get(`${basePath}`);
 
     expect(response.statusCode).toBe(500);
     expect(response.body).toBeTruthy();
